🚚 Move emulator functions to a subdirectory
