+++ src/functions/emulator/start-spanner.spec.ts
@@ -18,9 +18,9 @@ import { Spanner } from '@google-cloud/spanner';
 import { credentials } from '@grpc/grpc-js';
 import { jest } from '@jest/globals';
 import 'jest-extended';
-import { EmulatorStartForSpanner } from './emulator-start-spanner.js';
-import { EmulatorStopForSpanner } from './emulator-stop-spanner.js';
-import { GoogleSpannerListDatabases } from './google-spanner-list-databases.js';
+import { GoogleSpannerListDatabases } from '../google-spanner-list-databases.js';
+import { EmulatorStartForSpanner } from './start-spanner.js';
+import { EmulatorStopForSpanner } from './stop-spanner.js';
 
 describe('EmulatorStartForSpanner', () => {
   let context: WorkspaceContext;
